# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Learning Management System (LMS) API
  description: |
    API for managing a Learning Management System with courses, modules, lessons, and user tracking.
    This API provides comprehensive functionality for creating and managing educational content,
    user enrollments, and progress tracking.
  version: 1.0.0
  contact:
    name: LMS API Support
    email: support@lmsapi.example.com
servers:
  - url: https://api.lms.example.com/v1
    description: Production server
  - url: https://staging-api.lms.example.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    Error:
      type: object
      properties:
        id:
          type: string
          example: api.course.get
        ver:
          type: string
          example: "1.0"
        ts:
          type: string
          format: date-time
        params:
          type: object
          properties:
            resmsgid:
              type: string
              format: uuid
            status:
              type: string
              enum: [failed]
            err:
              type: string
              example: RESOURCE_NOT_FOUND
            errmsg:
              type: string
              example: Course not found
        responseCode:
          type: integer
          example: 404
        result:
          type: object
...
    CourseTracking:
      type: object
      properties:
        courseTrackId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
          nullable: true
        noOfLessons:
          type: integer
        completedLessons:
          type: integer
        status:
          type: string
          enum: [incomplete, in_progress, completed]
        lastAccessedDate:
          type: string
          format: date-time
        certGenDate:
          type: string
          format: date-time
          nullable: true
      required:
        - courseId
        - userId
        - startDatetime
    CourseTrackingWithDetails:
      allOf:
        - $ref: '#/components/schemas/CourseTracking'
        - type: object
          properties:
            course:
              type: object
              properties:
                title:
                  type: string
                shortDescription:
                  type: string
                image:
                  type: string
            progress:
              type: integer
              description: Percentage of course completed
...
    SuccessResponse:
      type: object
      properties:
        id:
          type: string
        ver:
          type: string
        ts:
          type: string
          format: date-time
        params:
          type: object
          properties:
            resmsgid:
              type: string
              format: uuid
            status:
              type: string
              enum: [successful]
            err:
              type: object
            errmsg:
              type: string
        responseCode:
          type: integer
        result:
          type: object
...
security:
  - bearerAuth: []
paths:
...
  /courses/{courseId}/tracking:
    post:
      summary: Start course tracking
      description: Starts tracking a user's progress in a course
      operationId: startCourseTracking
      tags:
        - Course Tracking
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                startDatetime:
                  type: string
                  format: date-time
              required:
                - userId
                - startDatetime
      responses:
        '201':
          description: Course tracking started successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      responseCode:
                        type: integer
                        example: 201
                      result:
                        $ref: '#/components/schemas/CourseTracking'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update course tracking
      description: Updates a user's progress in a course
      operationId: updateCourseTracking
      tags:
        - Course Tracking
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                completedLessons:
                  type: integer
                status:
                  type: string
                  enum: [incomplete, in_progress, completed]
                lastAccessedDate:
                  type: string
                  format: date-time
              required:
                - userId
      responses:
        '200':
          description: Course tracking updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/CourseTracking'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course tracking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: Get course tracking
      description: Retrieves course tracking for a specific user and course
      operationId: getCourseTracking
      tags:
        - Course Tracking
      parameters:
        - name: courseId
          in: path
          required: true
          description: UUID of the course
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          description: UUID of the user
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/CourseTrackingWithDetails'
        '404':
          description: Course tracking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...